fixed game logic with guns. guns can only be used and created once

+++ src/Objects/Callbacks/GunCallback.java
@@ -3,5 +3,5 @@ package Objects.Callbacks;
 public interface GunCallback {
     void changeLED(int r, int g, int b);
 
-    void blink(int i, int r, int g, int b);
+    void blink(double i, int r, int g, int b);
 }
